requirements.txt
neat-python
gymnasium[box2d]
graphviz
matplotlib